requirements.txt
aiogram==3.24.0
python-dotenv==1.2.1
redis==5.0.0


